+++ library/src/main/ets/model/StormTableVersion.ts
@@ -1,9 +1,12 @@
 import { Column, SqlColumn, SqlTable, Table } from '../../../../Index'
 
+/**
+ * 用于记录每一个表的版本号信息
+ */
 class StormTableVersions extends Table<StormTableVersion> {
   override readonly tableName = 't_storm_table_version'
-  readonly name = Column.text('name')
-  readonly version = Column.integer('version')
+  readonly name = Column.text('table_name')
+  readonly version = Column.integer('table_version')
 }
 
 export const stormTableVersions = new StormTableVersions()